Decompile four of the functions recovered from the data segment
None of the forty-two functions rescued in 63bc7df3 had been decompiled, and
the smallest are only a few instructions each.

The debug printf is two instructions - it returns 1 and does nothing - and must
be defined without a prototype, since a variadic one makes gcc emit the
argument-saving prologue the retail function does not have.

Two details had to match the original rather than the obvious spelling: the pad
initialiser derives its second buffer as an offset from the first, which is why
retail computes it with one instruction instead of loading a second address;
and the draw-mode packet builder passes its tpage through as a full word, so a
u16 parameter inserted a truncation retail does not perform.

Figure moves to 1010/1106. ROM stays byte-identical (sha1 2913e156).

+++ include/game/render.h
@@ -612,7 +612,7 @@ u8 *GameQueueTexturedRect(
  * place (no AddPrim, no cursor advance) and has no callers in the retail EXE.
  */
 u8 *GameQueueDrawModePrim(void *ot, u8 *prim, u16 tpage) asm("func_80017390");
-void GameSetDrawModePacket(u8 *prim, u16 tpage) asm("func_80017760");
+void GameSetDrawModePacket(u8 *prim, s32 tpage) asm("func_80017760");
 
 /*
  * A third font, separate from the small (6x12) / large (8x16) pair above:
